try to fix reddit img fetching
if we pass a reddit url to the qml Image component, the component is
redirected to a reddit html image viewer and this breaks the image
component.
So we must download the image ourselves via JS, convert it to base64
and set the qml Image source to the base64 data uri.